fix(progress): change progress field from String to Integer (0-100)
- progress column: String(8) -> Integer
- Update model, service, and engine to use int instead of float/string
- Fixes "Data too long for column" error

Note: Database migration needed:
  ALTER TABLE backtest_task MODIFY COLUMN progress INT DEFAULT 0;

+++ src/ginkgo/trading/engines/time_controlled_engine.py
@@ -929,9 +929,9 @@ class TimeControlledEventEngine(EventEngine, ITimeAwareComponent):
                 elapsed_days = (current_time - start_time).days
 
                 if total_days > 0:
-                    progress = min(100.0, max(0.0, (elapsed_days / total_days) * 100))
+                    progress = int(min(100, max(0, (elapsed_days / total_days) * 100)))
                 else:
-                    progress = 100.0
+                    progress = 100
 
                 # 调用回调
                 self._progress_callback(progress, str(current_time.date()))
